updated SABase to make it 64-bit clean

+++ Foundation/NSArray+SA_Additions.m
@@ -106,8 +106,8 @@
 	NSUInteger						count = self.count;
 	
 	for (NSUInteger i = 0; i < count; ++i) {		// Select a random element between i and end of array to swap with.
-		int nElements = count - i;
-		int n = (arc4random() % nElements) + i;
+		NSUInteger nElements = count - i;
+		NSUInteger n = (arc4random() % nElements) + i;
 		[copy exchangeObjectAtIndex: i withObjectAtIndex: n];
 	}
